+++ artiq/frontend/artiq_browser.py
@@ -8,7 +8,7 @@ import logging
 import sys
 
 from PyQt5 import QtCore, QtGui, QtWidgets
-from quamash import QEventLoop
+from qasync import QEventLoop
 
 from sipyco.asyncio_tools import atexit_register_coroutine
 from sipyco import common_args
@@ -144,9 +144,6 @@ def main():
     app = QtWidgets.QApplication(["ARTIQ Browser"])
     loop = QEventLoop(app)
     asyncio.set_event_loop(loop)
-    # https://github.com/harvimt/quamash/issues/123
-    if sys.version_info >= (3, 8):
-        asyncio.events._set_running_loop(loop)
     atexit.register(loop.close)
 
     datasets_sub = models.LocalModelManager(datasets.Model)