+++ artiq/frontend/artiq_dashboard.py
@@ -8,7 +8,7 @@ import logging
 import sys
 
 from PyQt5 import QtCore, QtGui, QtWidgets
-from quamash import QEventLoop
+from qasync import QEventLoop
 
 from sipyco.pc_rpc import AsyncioClient, Client
 from sipyco.broadcast import Receiver
@@ -104,9 +104,6 @@ def main():
     app = QtWidgets.QApplication(["ARTIQ Dashboard"])
     loop = QEventLoop(app)
     asyncio.set_event_loop(loop)
-    # https://github.com/harvimt/quamash/issues/123
-    if sys.version_info >= (3, 8):
-        asyncio.events._set_running_loop(loop)
     atexit.register(loop.close)
     smgr = state.StateManager(args.db_file)